Rebuild for -no_rqs on Irix 5.3.
Also because tcl/tk is built with cc so must expect be or major makefile
hackery would be needed. As it is the hackerey is now in build.sh to make
sure the right params are used accordingly to how tcl was built.
The -no-rpath patch simply nukes any attempts to include an rpath with the
builddir in the finished binary.